Enable macOS (10.23) (#13)
* Allow macOS builds of this recipe

* MNT: Re-rendered with conda-smithy 2.4.0

Now that the `skip` has changes to allow macOS builds of the recipe,
re-render to enable Travis CI to build the recipe.

* Build with -O3 optimizations

Not only should this make pcre2 more performant, it appears to fix a
test failure that occurs on macOS.

ref: https://bugs.exim.org/show_bug.cgi?id=1642

* Bump build number to 2

Needed as we have changed the compiler optimizations used to build
pcre2. So this will alter the existing Linux build.